docs: remove stale data.ts files
