--- docs/src/app/(public)/(content)/react/components/autocomplete/demos/fuzzy-matching/css-modules/data.ts
@@ -1,83 +1,0 @@
-export interface FuzzyItem {
-  title: string;
-  description: string;
-  category: string;
-}
-
-export const fuzzyItems: FuzzyItem[] = [
-  {
-    title: 'React Hooks Guide',
-    description: 'Learn how to use React Hooks like useState, useEffect, and custom hooks',
-    category: 'React',
-  },
-  {
-    title: 'JavaScript Array Methods',
-    description: 'Master array methods like map, filter, reduce, and forEach in JavaScript',
-    category: 'JavaScript',
-  },
-  {
-    title: 'CSS Flexbox Layout',
-    description: 'Complete guide to CSS Flexbox for responsive web design',
-    category: 'CSS',
-  },
-  {
-    title: 'TypeScript Interfaces',
-    description: 'Understanding TypeScript interfaces and type definitions',
-    category: 'TypeScript',
-  },
-  {
-    title: 'React Performance Optimization',
-    description: 'Tips and techniques for optimizing React application performance',
-    category: 'React',
-  },
-  {
-    title: 'HTML Semantic Elements',
-    description: 'Using semantic HTML elements for better accessibility and SEO',
-    category: 'HTML',
-  },
-  {
-    title: 'Node.js Express Server',
-    description: 'Building RESTful APIs with Node.js and Express framework',
-    category: 'Node.js',
-  },
-  {
-    title: 'Vue Composition API',
-    description: 'Modern Vue.js development using the Composition API',
-    category: 'Vue.js',
-  },
-  {
-    title: 'Angular Components',
-    description: 'Creating reusable Angular components with TypeScript',
-    category: 'Angular',
-  },
-  {
-    title: 'Python Django Framework',
-    description: 'Web development with Python Django framework',
-    category: 'Python',
-  },
-  {
-    title: 'CSS Grid Layout',
-    description: 'Advanced CSS Grid techniques for complex layouts',
-    category: 'CSS',
-  },
-  {
-    title: 'React Testing Library',
-    description: 'Testing React components with React Testing Library',
-    category: 'React',
-  },
-  {
-    title: 'MongoDB Queries',
-    description: 'Advanced MongoDB queries and aggregation pipelines',
-    category: 'Database',
-  },
-  {
-    title: 'Webpack Configuration',
-    description: 'Optimizing Webpack configuration for production builds',
-    category: 'Build Tools',
-  },
-  {
-    title: 'SASS/SCSS Guide',
-    description: 'Writing maintainable CSS with SASS and SCSS',
-    category: 'CSS',
-  },
-];
