--- docs/src/app/(public)/(content)/react/components/autocomplete/demos/grouped/css-modules/data.ts
@@ -1,65 +1,0 @@
-export interface Tag {
-  id: string;
-  label: string;
-  group: 'Type' | 'Component';
-}
-
-export interface TagGroup {
-  value: string;
-  items: Tag[];
-}
-
-const tagsData: Tag[] = [
-  { id: 't1', label: 'feature', group: 'Type' },
-  { id: 't2', label: 'fix', group: 'Type' },
-  { id: 't3', label: 'bug', group: 'Type' },
-  { id: 't4', label: 'docs', group: 'Type' },
-  { id: 't5', label: 'internal', group: 'Type' },
-  { id: 't6', label: 'mobile', group: 'Type' },
-  { id: 'c-accordion', label: 'component: accordion', group: 'Component' },
-  { id: 'c-alert-dialog', label: 'component: alert dialog', group: 'Component' },
-  { id: 'c-autocomplete', label: 'component: autocomplete', group: 'Component' },
-  { id: 'c-avatar', label: 'component: avatar', group: 'Component' },
-  { id: 'c-checkbox', label: 'component: checkbox', group: 'Component' },
-  { id: 'c-checkbox-group', label: 'component: checkbox group', group: 'Component' },
-  { id: 'c-collapsible', label: 'component: collapsible', group: 'Component' },
-  { id: 'c-combobox', label: 'component: combobox', group: 'Component' },
-  { id: 'c-context-menu', label: 'component: context menu', group: 'Component' },
-  { id: 'c-dialog', label: 'component: dialog', group: 'Component' },
-  { id: 'c-field', label: 'component: field', group: 'Component' },
-  { id: 'c-fieldset', label: 'component: fieldset', group: 'Component' },
-  { id: 'c-filterable-menu', label: 'component: filterable menu', group: 'Component' },
-  { id: 'c-form', label: 'component: form', group: 'Component' },
-  { id: 'c-input', label: 'component: input', group: 'Component' },
-  { id: 'c-menu', label: 'component: menu', group: 'Component' },
-  { id: 'c-menubar', label: 'component: menubar', group: 'Component' },
-  { id: 'c-meter', label: 'component: meter', group: 'Component' },
-  { id: 'c-navigation-menu', label: 'component: navigation menu', group: 'Component' },
-  { id: 'c-number-field', label: 'component: number field', group: 'Component' },
-  { id: 'c-popover', label: 'component: popover', group: 'Component' },
-  { id: 'c-preview-card', label: 'component: preview card', group: 'Component' },
-  { id: 'c-progress', label: 'component: progress', group: 'Component' },
-  { id: 'c-radio', label: 'component: radio', group: 'Component' },
-  { id: 'c-scroll-area', label: 'component: scroll area', group: 'Component' },
-  { id: 'c-select', label: 'component: select', group: 'Component' },
-  { id: 'c-separator', label: 'component: separator', group: 'Component' },
-  { id: 'c-slider', label: 'component: slider', group: 'Component' },
-  { id: 'c-switch', label: 'component: switch', group: 'Component' },
-  { id: 'c-tabs', label: 'component: tabs', group: 'Component' },
-  { id: 'c-toast', label: 'component: toast', group: 'Component' },
-  { id: 'c-toggle', label: 'component: toggle', group: 'Component' },
-  { id: 'c-toggle-group', label: 'component: toggle group', group: 'Component' },
-  { id: 'c-toolbar', label: 'component: toolbar', group: 'Component' },
-  { id: 'c-tooltip', label: 'component: tooltip', group: 'Component' },
-];
-
-function groupTags(tags: Tag[]): TagGroup[] {
-  const groups: { [key: string]: Tag[] } = {};
-  tags.forEach((t) => {
-    (groups[t.group] ??= []).push(t);
-  });
-  const order = ['Type', 'Component'];
-  return order.map((value) => ({ value, items: groups[value] ?? [] }));
-}
-
-export const groupedTags = groupTags(tagsData);
