--- docs/src/app/(public)/(content)/react/components/autocomplete/demos/inline/css-modules/data.ts
@@ -1,48 +1,0 @@
-export interface Tag {
-  id: string;
-  label: string;
-}
-
-export const tags: Tag[] = [
-  { id: 't1', label: 'feature' },
-  { id: 't2', label: 'fix' },
-  { id: 't3', label: 'bug' },
-  { id: 't4', label: 'docs' },
-  { id: 't5', label: 'internal' },
-  { id: 't6', label: 'mobile' },
-  { id: 'c-accordion', label: 'component: accordion' },
-  { id: 'c-alert-dialog', label: 'component: alert dialog' },
-  { id: 'c-autocomplete', label: 'component: autocomplete' },
-  { id: 'c-avatar', label: 'component: avatar' },
-  { id: 'c-checkbox', label: 'component: checkbox' },
-  { id: 'c-checkbox-group', label: 'component: checkbox group' },
-  { id: 'c-collapsible', label: 'component: collapsible' },
-  { id: 'c-combobox', label: 'component: combobox' },
-  { id: 'c-context-menu', label: 'component: context menu' },
-  { id: 'c-dialog', label: 'component: dialog' },
-  { id: 'c-field', label: 'component: field' },
-  { id: 'c-fieldset', label: 'component: fieldset' },
-  { id: 'c-filterable-menu', label: 'component: filterable menu' },
-  { id: 'c-form', label: 'component: form' },
-  { id: 'c-input', label: 'component: input' },
-  { id: 'c-menu', label: 'component: menu' },
-  { id: 'c-menubar', label: 'component: menubar' },
-  { id: 'c-meter', label: 'component: meter' },
-  { id: 'c-navigation-menu', label: 'component: navigation menu' },
-  { id: 'c-number-field', label: 'component: number field' },
-  { id: 'c-popover', label: 'component: popover' },
-  { id: 'c-preview-card', label: 'component: preview card' },
-  { id: 'c-progress', label: 'component: progress' },
-  { id: 'c-radio', label: 'component: radio' },
-  { id: 'c-scroll-area', label: 'component: scroll area' },
-  { id: 'c-select', label: 'component: select' },
-  { id: 'c-separator', label: 'component: separator' },
-  { id: 'c-slider', label: 'component: slider' },
-  { id: 'c-switch', label: 'component: switch' },
-  { id: 'c-tabs', label: 'component: tabs' },
-  { id: 'c-toast', label: 'component: toast' },
-  { id: 'c-toggle', label: 'component: toggle' },
-  { id: 'c-toggle-group', label: 'component: toggle group' },
-  { id: 'c-toolbar', label: 'component: toolbar' },
-  { id: 'c-tooltip', label: 'component: tooltip' },
-];
