--- docs/src/app/(public)/(content)/react/components/combobox/demos/creatable/css-modules/data.ts
@@ -1,13 +1,0 @@
-export interface LabelItem {
-  creatable?: string;
-  id: string;
-  value: string;
-}
-
-export const initialLabels: LabelItem[] = [
-  { id: 'bug', value: 'bug' },
-  { id: 'docs', value: 'documentation' },
-  { id: 'enhancement', value: 'enhancement' },
-  { id: 'help-wanted', value: 'help wanted' },
-  { id: 'good-first-issue', value: 'good first issue' },
-];
