--- docs/src/app/(public)/(content)/react/components/combobox/demos/hero/css-modules/data.ts
@@ -1,1 +1,0 @@
-export const options = ['Apple', 'Banana', 'Orange', 'Pineapple', 'Grape', 'Mango', 'Strawberry'];
