--- docs/src/app/(public)/(content)/react/components/combobox/demos/multiple/css-modules/data.ts
@@ -1,18 +1,0 @@
-export interface Lang {
-  id: string;
-  value: string;
-}
-
-export const langs: Lang[] = [
-  { id: 'js', value: 'JavaScript' },
-  { id: 'ts', value: 'TypeScript' },
-  { id: 'py', value: 'Python' },
-  { id: 'java', value: 'Java' },
-  { id: 'cpp', value: 'C++' },
-  { id: 'cs', value: 'C#' },
-  { id: 'php', value: 'PHP' },
-  { id: 'ruby', value: 'Ruby' },
-  { id: 'go', value: 'Go' },
-  { id: 'rust', value: 'Rust' },
-  { id: 'swift', value: 'Swift' },
-];
