--- docs/src/app/(public)/(content)/react/components/combobox/demos/virtualized/css-modules/data.ts
@@ -1,4 +1,0 @@
-export const virtualItems = Array.from({ length: 10000 }, (_, i) => {
-  const indexLabel = String(i + 1).padStart(4, '0');
-  return `Item ${indexLabel}`;
-});
